Updated project structure
The commit takes care of the extensibility problem with the traditional packaging of the software. With the new structure, the program is optimised for the installation of new modules and functionality. A ModuleManager app has been added to take care of the available and installed modules in the system. The salary module has been separated as a standalone module. the program is divided into the 5 "base" modules and packages for the extension modules, such as Salary. A mongo.js file has been added to configure the MongoDB database on the host machine. Optimisations on the applySchedule() method.